# docs/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "vocascan",
    "description": "A highly configurable vocabulary trainer.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "description": "current",
      "url": "/api"
    },
    {
      "description": "localhost",
      "url": "http://localhost:5000/api"
    }
  ],
...
  "paths": {
...
    "/languagePackage/{languagePackageId}/group": {
      "post": {
        "summary": "Create group for language package",
        "tags": ["group"],
        "operationId": "createGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/languagePackageId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateGroup"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CreateGroup"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "get": {
        "summary": "Get groups belonging to language package",
        "tags": ["group"],
        "operationId": "getGroups",
        "parameters": [
          {
            "$ref": "#/components/parameters/languagePackageId"
          },
          {
            "$ref": "#/components/parameters/onlyActivated"
          },
          {
            "$ref": "#/components/parameters/onlyStaged"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetGroups"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
    "/group/{groupId}/vocabulary": {
      "get": {
        "summary": "Get all vocabulary belonging to specific group",
        "tags": ["vocabulary"],
        "operationId": "getGroupVocabulary",
        "parameters": [
          {
            "$ref": "#/components/parameters/groupId"
          },
          {
            "$ref": "#/components/parameters/search"
          },
          {
            "$ref": "#/components/parameters/onlyStaged"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetGroupVocabulary"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
    "/vocabulary/{vocabularyId}": {
      "delete": {
        "summary": "Delete vocabulary card",
        "tags": ["vocabulary"],
        "operationId": "deleteVocabulary",
        "parameters": [
          {
            "$ref": "#/components/parameters/vocabularyId"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "put": {
        "summary": "Modify vocabulary card",
        "tags": ["vocabulary"],
        "operationId": "modifyVocabulary",
        "parameters": [
          {
            "$ref": "#/components/parameters/vocabularyId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ModifyVocabulary"
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "summary": "Handles your result. moves vocabulary in the drawers depending on whether they were answered correctly or incorrectly",
        "tags": ["query"],
        "operationId": "checkVocabulary",
        "parameters": [
          {
            "$ref": "#/components/parameters/vocabularyId"
          },
          {
            "$ref": "#/components/parameters/answer"
          },
          {
            "$ref": "#/components/parameters/progress"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/CheckQuery"
          },
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/group/{groupId}": {
      "delete": {
        "summary": "Delete group",
        "tags": ["group"],
        "operationId": "deleteGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/groupId"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "put": {
        "summary": "Modify group",
        "tags": ["group"],
        "operationId": "modifyGroup",
        "parameters": [
          {
            "$ref": "#/components/parameters/groupId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ModifyGroup"
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Group": {
        "description": "Group",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "languagePackageId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "format": "uuid"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "Vocabulary": {
        "description": "Vocabulary",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "translations": {
            "type": "array",
            "items": {
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Progress": {
        "description": "Progress",
        "type": "object",
        "properties": {
          "queryProgress": {
            "type": "object",
            "properties": {
              "dueToday": {
                "type": "string"
              },
              "correct": {
                "type": "string"
              },
              "wrong": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "parameters": {
      "languagePackageId": {
        "in": "path",
        "name": "languagePackageId",
        "description": "Id from language package",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "groupId": {
        "in": "path",
        "name": "groupId",
        "description": "Id from group",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "vocabularyId": {
        "in": "path",
        "name": "vocabularyId",
        "description": "Id from vocabulary card",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "search": {
        "in": "query",
        "name": "search",
        "description": "Filter vocabulary by names and translation names",
        "schema": {
          "type": "string"
        }
      },
...
      "onlyActivated": {
        "in": "query",
        "name": "onlyActivated",
        "description": "Only return packages/groups/vocabs if it has at least one activated vocab card",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "onlyStaged": {
        "in": "query",
        "name": "onlyStaged",
        "description": "Only return packages/groups/vocabs if it hast at least one unactivated vocab card",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
      "answer": {
        "in": "query",
        "name": "answer",
        "description": "type in whether your answer was right or wrong",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "progress": {
        "in": "query",
        "name": "progress",
        "description": "If true, the progress will be returned",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
    },
    "requestBodies": {
...
      "CreateGroup": {
        "description": "Create group details",
        "content": {
          "application/json": {
            "schema": {
              "description": "Create group details",
              "type": "object",
              "required": ["name", "active"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "active": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "ModifyGroup": {
        "description": "Modify group details",
        "content": {
          "application/json": {
            "schema": {
              "description": "Modify group details",
              "type": "object",
              "required": ["name", "active"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "active": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
...
      "ModifyVocabulary": {
        "description": "add foreign word and translations",
        "content": {
          "application/json": {
            "schema": {
              "description": "Create vocabulary details",
              "type": "object",
              "required": ["name", "active"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "active": {
                  "type": "boolean"
                },
                "translations": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "CreateGroup": {
        "description": "Create group for language package",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Group"
            }
          }
        }
      },
...
      "GetGroups": {
        "description": "Returns info about Groups belonging to a language package",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Group"
              }
            }
          }
        }
      },
      "GetGroupVocabulary": {
        "description": "Returns info about all Vocabulary belonging to a group",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Vocabulary"
              }
            }
          }
        }
      },
...
      "CheckQuery": {
        "description": "If progress = true: Progress gets returned",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Progress"
            }
          }
        }
      },
...
      "204": {
        "description": "APIEmpty is an empty response"
      },
      "400": {
        "description": "APIClientError is a error format response"
      },
      "401": {
        "description": "APIUnauthorizedError is a not authorized error response"
      },
...
      "404": {
        "description": "APINotFound is a not found error response"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearerAuthAdmin": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Your account needs needs admin rights"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}